requirements.txt
Flask
Flask-Cors
Flask-PyMongo
python-dotenv
pymongo
requests
Flask-JWT-Extended